+++ CHANGES.rst
@@ -47,6 +47,7 @@ ChangeLog
 * Split-up ``supported_territory_codes()`` into ``supported_country_codes()``
   and ``supported_subdivision_codes()``.
 
+
 0.4.3 (2015-03-26)
 ------------------