+++ CHANGES.rst
@@ -2,7 +2,7 @@ ChangeLog
 =========
 
 
-0.6.2 (unreleased)
+0.7.0 (unreleased)
 ------------------
 
 * Include internal fields disguised as properties when ``repr``-ing the